Reorganized, added doc strings and type hints.
Added doc strings and type hints to nearly all
functions.

Reorganized the folder and file structure to
make navigating through the application project
more intuitive.

Modularized the project a bit more to reduce
coupling.

Removed many of the images that are actually
present in the sprite sheets and tile sheets,
but which were only present because they were
useful during development for creating a tiled
map.

Replaced some miscellaneous classes with class
methods. For example, instead of a BigTank class,
the Tank class now has a method for creating a
big tank.